no changes made

+++ src/main/java/com/githubTheTrueHooha/springgraphql/Repository/EmployeeRepository.java
@@ -3,6 +3,8 @@ package com.githubTheTrueHooha.springgraphql.Repository;
 import com.githubTheTrueHooha.springgraphql.Model.Employees;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employees, Long> {
+public interface EmployeeRepository extends JpaRepository<Employees, Integer> {
+
+
 
 }
